+++ fleetpings/__init__.py
@@ -2,5 +2,5 @@
 a couple of variable to use throughout the app
 """
 
-__version__ = "2.12.0"
+__version__ = "2.13.0-beta.1"
 __title__ = "Fleet Pings"